TMP: Convert metadata update test
